requirements.txt
# Database and ORM
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Environment and configuration
python-dotenv==1.0.0

# Email services
sendgrid==6.10.0

# AI/OpenAI
openai==1.3.7

# Web framework (for API endpoints)
fastapi==0.104.1
uvicorn==0.24.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.5.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 